Refactor and create FactomWalletdKeyStore class
