+++ test/app/index.spec.js
@@ -4,6 +4,7 @@ const app = require('../../src/app');
 describe('Factom identity for applications module exports', function () {
 
     it('Should export app module', function () {
+        assert.isFunction(app.FactomWalletdKeyStore);
         assert.isFunction(app.FactomIdentityManager);
         assert.isFunction(app.isValidIdentityKey);
         assert.isFunction(app.isValidPublicIdentityKey);
